+++ entities/player_test.go
@@ -39,8 +39,8 @@ func TestCreateMission(t *testing.T) {
 		Color:          Color{22, 22, 22},
 		TwitterID:      "asdf",
 		HomePlanet:     "planet.271_203",
-		ScreenSize:     []int{1, 1},
-		ScreenPosition: []int{2, 2},
+		ScreenSize:     []uint16{1, 1},
+		ScreenPosition: []int64{2, 2},
 	}
 
 	validMission := player.StartMission(&planetStart, &planetEnd, 80, "Attack")
